feat: changed long to int on Product model

+++ API/Models/Product.cs
@@ -2,7 +2,7 @@
 {
     public class Product
     {
-        public long id { get; set; } = -1;
+        public int id { get; set; } = -1;
         public string name { get; set; } = "";
         public decimal price { get; set; } = -1;
     }
@@ -13,7 +13,7 @@
     }
     public class ProductUpdate
     {
-        public long id { get; set; } = -1;
+        public int id { get; set; } = -1;
         public string name { get; set; } = "";
         public decimal price { get; set; } = -1;
     }
